Refactor src
